Update and rename README.md to  LED Blinking Using Arduino
Blinking an LED (  c++ codes included)

Blinking an LED is an introductory Arduino project in which we control an LED using Arduino. LED blinking refers to the process of continuously turning an LED (Light Emitting Diode) and off in a repetitive pattern. It is a simple and common demonstration in electronics and microcontroller-based projects.
Working Procedure

setup() and loop() are two fundamental Arduino functions for controlling the behavior of your board. The Arduino framework automatically calls these functions, which form the foundation of any Arduino program.

The setup() function is only called once when the Arduino board boots up or is reset. Its goal is to set pin modes, initialize variables, and execute any other necessary setup tasks before the main loop begins. This function can be used to configure settings that should only be changed once over the board’s lifespan.

The loop() function is the heart of an Arduino program. After the setup() function is executed, the loop() function starts running repeatedly until the Arduino is powered off or reset. It contains the main code that performs the desired tasks, controls the board, user input. Whatever is included in the loop() function will be executed in a continuous loop, allowing the Arduino to perform its intended functions continuously.

In the code, we have declared two integers, LEDpin and delayT. LEDpin represents the pin number of the Arduino where LEDs need to be connected, and delayT is an integer variable for the delay() function. The delay() function accepts values in milliseconds. 


Components Required

    1 X LED
    1 X Resistor, 330 Ohm
    Breadboard
    Arduino UNO R4 or earlier versions.
    Jumper wires

Components Description

    1 X LED: We are controlling only one LED in this program.
    1 X Resistor, 330 Ohm: For every LED, we need one current limiting resistor.
    Breadboard: A breadboard is a fundamental tool used in electronics and prototyping to build and test circuits without soldering.
    Arduino UNO R4 or earlier versions.
    Jumper wires: Jumper wires are simple electrical wires with connectors on both ends used to create connections between various electronic components or points on a circuit on a breadboard.
 for more information about led light panels and channelumes just wisit us at https://kordisign.com/

light box = https://kordisign.com/light-box
neons = https://kordisign.com/%d8%aa%d8%a7%d8%a8%d9%84%d9%88-%d9%86%d8%a6%d9%88%d9%86/